Move actionreplay to tests
