+++ triforce_lib/__init__.py
@@ -7,7 +7,6 @@ import json
 import retro
 
 from .zelda_orchestrator import ZeldaAIOrchestrator
-from .action_replay import ZeldaActionReplay
 from .zeldaml import ZeldaML
 from .scenario import ZeldaScenario
 from .models import ZeldaModel
@@ -22,7 +21,6 @@ __all__ = [
     ZeldaML.__name__,
     ZeldaModel.__name__,
     ZeldaScenario.__name__,
-    ZeldaActionReplay.__name__,
     ZeldaAIOrchestrator.__name__,
     is_in_cave.__name__,
     ]
